% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gray_scale.R
\name{filter_gray_scale}
\alias{filter_gray_scale}
\title{filter_gray_scale
Convert a given RGB pixel to grayscale.}
\usage{
filter_gray_scale(pixel_rgb_numeric)
}
\arguments{
\item{pixel_rgb_numeric}{A numeric vector containing RGB values (not raw).}
}
\value{
A raw vector with the modified RGB values.
}
\description{
This function converts a color pixel to grayscale using a standard
formula for luminance: `0.299*R + 0.587*G + 0.114*B`.
}
