% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_pixels_function.R
\name{replace_pixels}
\alias{replace_pixels}
\title{Replace Pixels with a Filter Function}
\usage{
replace_pixels(bitmap, filter_function = filter_gray_scale, ...)
}
\arguments{
\item{bitmap}{A bitmap image (e.g., from the magick package).}

\item{filter_function}{A transformation function to apply to each pixel (e.g., `filter_gray_scale` or `filter_two_color`).}

\item{...}{Additional arguments passed to `filter_function` (e.g., `cutoff` for `filter_two_color`).}
}
\value{
A filtered bitmap with modified RGB values.
}
\description{
This function applies a pixel-level transformation to a bitmap image using a specified filter function.
The filter function can be `filter_gray_scale`, `filter_two_color`, or any custom filter function that returns a raw RGB triplet
}
